ITS#9012 one more Win64 fix
Was omitted from dda779d83e701c43ef48780653f396329e16c755

+++ servers/slapd/config.c
@@ -133,7 +133,7 @@ int config_check_vals(ConfigTable *Conf, ConfigArgs *c, int check_only ) {
 	int rc, arg_user, arg_type, arg_syn, iarg;
 	unsigned uiarg;
 	long larg;
-	unsigned long ularg;
+	size_t ularg;
 	ber_len_t barg;
 	
 	if(Conf->arg_type == ARG_IGNORED) {
